requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
databases==0.9.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.1
langchain-core==1.0.0
langchain-openai==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.4
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
passlib==1.7.4
pgvector==0.4.1
psycopg2-binary==2.9.9
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
langchain-openai
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
apscheduler